# Improve index entries deduping and concurrency